Move mklcpu implementation to mkl_common
